+++ server/db.json
@@ -91,6 +91,15 @@
         "price": 29000,
         "imageUrl": "https://cdn-mart.baemin.com/sellergoods/main/6b95c66a-c13d-4ccd-9df5-b1af1428a225.jpg"
       }
+    },
+    {
+      "id": 1645332480590,
+      "product": {
+        "id": 3,
+        "name": "펩시 콜라 355ml 24캔",
+        "price": 83700,
+        "imageUrl": "https://cdn-mart.baemin.com/sellergoods/main/84fc0238-0239-4d0e-870b-a9daa6f2c42c.jpg"
+      }
     }
   ],
   "orders": [
